[REFACTOR] #202 ProfileView extension 분리
